Добавляет на страницу точку входа скриптов
