+++ js/utils.js
@@ -1,0 +1,4 @@
+// Функция для нахождения случайного числа
+const getRandomInteger = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
+
+export {getRandomInteger};
